+++ src/rdkafka_op.h
@@ -181,6 +181,10 @@ typedef enum {
                                                      u.admin_request >*/
         RD_KAFKA_OP_LISTOFFSETS,     /**< Admin: ListOffsets u.admin_request >*/
         RD_KAFKA_OP_METADATA_UPDATE, /**< Metadata update (KIP 951) **/
+        RD_KAFKA_OP_SET_TELEMETRY_BROKER, /**< Set preferred broker for
+                                               telemetry. */
+        RD_KAFKA_OP_TERMINATE_TELEMETRY,  /**< Start termination sequence for
+                                               telemetry. */
         RD_KAFKA_OP__END
 } rd_kafka_op_type_t;
 
@@ -573,6 +577,8 @@ struct rd_kafka_op_s {
                                RD_KAFKA_MOCK_CMD_BROKER_SET_RACK,
                                RD_KAFKA_MOCK_CMD_COORD_SET,
                                RD_KAFKA_MOCK_CMD_APIVERSION_SET,
+                               RD_KAFKA_MOCK_CMD_REQUESTED_METRICS_SET,
+                               RD_KAFKA_MOCK_CMD_TELEMETRY_PUSH_INTERVAL_SET,
                         } cmd;
 
                         rd_kafka_resp_err_t err; /**< Error for:
@@ -612,6 +618,8 @@ struct rd_kafka_op_s {
                                                   *    TOPIC_CREATE (repl fact)
                                                   *    PART_SET_FOLLOWER_WMARKS
                                                   *    APIVERSION_SET (maxver)
+                                                  *    REQUESTED_METRICS_SET (metrics_cnt)
+                                                  *    TELEMETRY_PUSH_INTERVAL_SET (interval)
                                                   */
                         int32_t leader_id;       /**< Leader id, for:
                                                   *   PART_PUSH_LEADER_RESPONSE
@@ -619,6 +627,8 @@ struct rd_kafka_op_s {
                         int32_t leader_epoch;    /**< Leader epoch, for:
                                                   *   PART_PUSH_LEADER_RESPONSE
                                                   */
+                        char **metrics;          /**< Metrics requested, for:
+                                                  *   REQUESTED_METRICS_SET */
                 } mock;
 
                 struct {
@@ -680,6 +690,11 @@ struct rd_kafka_op_s {
                         void *opaque;
 
                 } leaders;
+
+                struct {
+                        /** Preferred broker for telemetry. */
+                        rd_kafka_broker_t *rkb;
+                } telemetry_broker;
 
         } rko_u;
 };